+++ src/messages/en/admin.php
@@ -437,4 +437,17 @@ return [
     'active_button_togglestatus_enabled' => '{modelName} enabled',
     'active_button_togglestatus_disabled' => '{modelName} disabled',
     'layout_filemanager_detail_details' => 'File details',
+
+// 2.3.0
+
+    'aws_delete_remove_success' => 'The tag and its relations have been removed.',
+    'aws_delete_remove_wrong_name' => 'The tag name is wrong.',
+    'aws_delete_relations_info' => 'The tag <b>{name}</b> is used <b>{count}</b> times. It is assigned to entries in the following tables:',
+    'aws_delete_relations_table_name' => 'Relation Table Name',
+    'aws_delete_relations_table_count' => 'Number of Entries',
+    'aws_delete_relations_form_hint' => 'In order to confirm the tag deletion, enter the name of the tag and press \"Delete\".',
+
+// 2.4.0
+
+    'aws_missing_callback_param_generic_errror' => 'Some required data is missing.',
 ];